File copy.

+++ module_renamer.py
@@ -1,5 +1,6 @@
 import os
 import shutil
+import re
 
 class ModuleRenamer(object):
 
@@ -17,23 +18,21 @@ class ModuleRenamer(object):
     # @todo Implement.
     return True
 
-  def rename_file_names(self):
+  def rename(self):
     new_module_dir = os.path.join(self.path_to_module, '..', self.to_name)
 
-    if __debug__:
-      # @todo remove after project complete
-      if os.path.exists(new_module_dir):
-        shutil.rmtree(new_module_dir)
+    if __debug__ and os.path.exists(new_module_dir):
+      shutil.rmtree(new_module_dir)
 
-    os.mkdir(new_module_dir)
+    shutil.copytree(self.path_to_module, new_module_dir)
 
-    for dirpath, dirnames, filenames in os.walk(self.path_to_module):
-
-      for subdirname in dirnames:
-        os.mkdir(os.path.join(self.path_to_module, '..', self.to_name, subdirname))
-
+    for dirpath, dirnames, filenames in os.walk(new_module_dir):
       for filename in filenames:
-        print('File: ' + os.path.join(dirpath, filename))
+        new_filename = re.sub(self.from_name, self.to_name, filename)
+        file_path = os.path.join(dirpath, filename)
+        new_file_path = os.path.join(dirpath, new_filename)
+        shutil.move(file_path, new_file_path)
+        self.rename_strings(new_file_path)
 
-  def rename_strings(self):
+  def rename_strings(self, filename):
     0
